Update settings to extract out to each file
