+++ admin_setting_cost_estimation.go
@@ -1,0 +1,80 @@
+package tfe
+
+import (
+	"context"
+)
+
+// Compile-time proof of interface implementation.
+var _ CostEstimationSettings = (*adminCostEstimationSettings)(nil)
+
+// CostEstimationSettings describes all the cost estimation admin settings.
+type CostEstimationSettings interface {
+	// Read returns the cost estimation settings.
+	Read(ctx context.Context) (*AdminCostEstimationSetting, error)
+
+	// Update updates the cost estimation settings.
+	Update(ctx context.Context, options AdminCostEstimationSettingOptions) (*AdminCostEstimationSetting, error)
+}
+
+type adminCostEstimationSettings struct {
+	client *Client
+}
+
+// AdminCostEstimationSetting represents the admin cost estimation settings.
+type AdminCostEstimationSetting struct {
+	ID                  string `jsonapi:"primary,cost-estimation-settings"`
+	Enabled             bool   `jsonapi:"attr,enabled"`
+	AWSAccessKeyID      string `jsonapi:"attr,aws-access-key-id"`
+	AWSAccessKey        string `jsonapi:"attr,aws-secret-key"`
+	GCPCredentials      string `jsonapi:"attr,gcp-credentials"`
+	AzureClientID       string `jsonapi:"attr,azure-client-id"`
+	AzureClientSecret   string `jsonapi:"attr,azure-client-secret"`
+	AzureSubscriptionID string `jsonapi:"attr,azure-subscription-id"`
+	AzureTenantID       string `jsonapi:"attr,azure-tenant-id"`
+}
+
+// Read returns the cost estimation settings.
+func (s *adminCostEstimationSettings) Read(ctx context.Context) (*AdminCostEstimationSetting, error) {
+	req, err := s.client.newRequest("GET", "admin/cost-estimation-settings", nil)
+	if err != nil {
+		return nil, err
+	}
+
+	ags := &AdminCostEstimationSetting{}
+	err = s.client.do(ctx, req, ags)
+	if err != nil {
+		return nil, err
+	}
+
+	return ags, nil
+}
+
+// AdminCostEstimationSettingOptions represents the admin options for updating
+// the cost estimation settings.
+// https://www.terraform.io/docs/cloud/api/admin/settings.html#request-body-1
+type AdminCostEstimationSettingOptions struct {
+	Enabled             *bool   `jsonapi:"attr,enabled,omitempty"`
+	AWSAccessKeyID      *string `jsonapi:"attr,aws-access-key-id,omitempty"`
+	AWSAccessKey        *string `jsonapi:"attr,aws-secret-key,omitempty"`
+	GCPCredentials      *string `jsonapi:"attr,gcp-credentials,omitempty"`
+	AzureClientID       *string `jsonapi:"attr,azure-client-id,omitempty"`
+	AzureClientSecret   *string `jsonapi:"attr,azure-client-secret,omitempty"`
+	AzureSubscriptionID *string `jsonapi:"attr,azure-subscription-id,omitempty"`
+	AzureTenantID       *string `jsonapi:"attr,azure-tenant-id,omitempty"`
+}
+
+// Update updates the cost-estimation settings.
+func (s *adminCostEstimationSettings) Update(ctx context.Context, options AdminCostEstimationSettingOptions) (*AdminCostEstimationSetting, error) {
+	req, err := s.client.newRequest("PATCH", "admin/cost-estimation-settings", &options)
+	if err != nil {
+		return nil, err
+	}
+
+	ace := &AdminCostEstimationSetting{}
+	err = s.client.do(ctx, req, ace)
+	if err != nil {
+		return nil, err
+	}
+
+	return ace, nil
+}
